can now switch between maps
Can now switch between maps in a 2 way fashion. not 1 way.

Player is drawn ,animated and collision logic is accounted for more logically now that we are using the current instance of the player

Cooldown for switching between maps and the player does not switch when standing in a gate.

Rewrote comments to make more sense

Removed all tile-based movement for the player as tile-based movement will not be used for this project.

draws hitboxes around gates as well