Added Evolution
Currently only supports pokemon who evolve by level.

planning on a descriptor language of sorts to store all the specific conditions and stuff in the games, as there tends to be a lot of that in pokemon.

that should help with different types of evolutions like friendships, stones, times of day etc etc.

also plan on finishing experience/levelling up